fixed confilcts

+++ gameserver/src/main/java/bomber/gameservice/controller/GameThread.java
@@ -17,7 +17,7 @@ public class GameThread implements Runnable {
     @Override
     public void run() {
         log.info("Start new thread called game-mechanics with gameId = " + gameId);
-        GameSession gameSession = new GameSession(gameId);
+        GameSession gameSession = new GameSession((int) gameId);
         log.info("Game has been init gameId={}", gameId);
         gameSessionMap.put(gameId, gameSession);
     }
